feat: enhance vault CLI commands with new functionalities and error handling
