+++ index.ts
@@ -1,6 +1,9 @@
 #! /usr/bin/env node
 
-import { program } from "./command/index.js";
-import "./programs/index.js";
+import { program } from './command/index.js';
+import './programs/index.js';
 
-program.parse(process.argv);
+program.parseAsync(process.argv).catch((error) => {
+  console.error('CLI Error:', error);
+  process.exit(1);
+});
