+++ programs/index.ts
@@ -1,5 +1,4 @@
 export * from './vault-hub.js';
-export * from './vault.js';
 export * from './vault-factory.js';
 export * from './dashboard.js';
 export * from './delegation.js';
@@ -7,3 +6,4 @@ export * from './voting.js';
 export * from './account.js';
 export * from './vault-hub-viewer.js';
 export * from './predeposit-guarantee.js';
+export * from './vault/index.js';
