+++ programs/vault/index.ts
@@ -1,0 +1,3 @@
+export * from './main.js';
+export * from './view.js';
+export * from './write.js';
