+++ programs/vault/main.ts
@@ -1,0 +1,3 @@
+import { program } from 'command';
+
+export const vault = program.command('vault').description('vault contract');
